requirements.txt
# /Users/mac/projects/ticket_bot/requirements.txt
aiogram==3.4.1
SQLAlchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.1
apscheduler==3.10.4
aiosqlite==0.19.0
